+++ packages/storefront-next-runtime/src/data-store/index.ts
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2026 Salesforce, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export { createDataStoreMiddleware } from './utils';
+export { createDataStoreContext } from './utils';
+export {
+    customSitePreferencesMiddleware,
+    DEFAULT_SITE_PREFERENCES_KEY,
+    getSitePreferences,
+    sitePreferencesContext,
+} from './middleware/custom-site-preferences';
+export {
+    customGlobalPreferencesContext,
+    customGlobalPreferencesMiddleware,
+    DEFAULT_CUSTOM_GLOBAL_PREFERENCES_KEY,
+    getCustomGlobalPreferences,
+} from './middleware/custom-global-preferences';
+export type { DataStoreMiddlewareOptions } from './utils';
+export type { SitePreferences } from './middleware/custom-site-preferences';
+export type { DataStoreContextKey, DataStoreEntryKey } from './utils';
+export type { DataStoreEntry, DataStoreProvider } from './provider';
+export type { CustomGlobalPreferences } from './middleware/custom-global-preferences';