+++ packages/storefront-next-dev/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.4.0-dev (Apr 07, 2026)
 
+- Add local dev support to MRT Data Layer [#1215](https://github.com/commerce-emu/storefront-next/pull/1215)
 - Clean up RSC/React Server Components references from documentation and comments ([#1363](https://github.com/commerce-emu/storefront-next/pull/1363))
 - Remove internal "odyssey" codename references from codebase
 - Add `sfnext scapi` CLI used to support generating and managing custom API clients in template projects (@W-21549425)