+++ packages/storefront-next-runtime/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v0.4.0-dev (Apr 07, 2026)
 
+- Add support to MRT Data Layer access [#1215](https://github.com/commerce-emu/storefront-next/pull/1215)
 - Add currency detection to site-context middleware (@W-21787262) ([#1342](https://github.com/commerce-emu/storefront-next/pull/1342))
 - Add runtime support for custom API client typing and proxy client composition (@W-21549425)